flask开发/聊天室/backend/requirements.txt
# Flask 核心框架
Flask==2.3.0
Flask-SQLAlchemy==3.0.5
Flask-SocketIO==5.3.4
Flask-CORS==4.0.0

# 数据库
PyMySQL==1.1.0
cryptography==41.0.3

# WebSocket
python-socketio==5.9.0
python-engineio==4.7.1

# 认证和加密
PyJWT==2.8.0
bcrypt==4.0.1

# 工具库
python-dotenv==1.0.0

# WSGI 服务器
eventlet==0.33.3
